:test_tube: :: (#4) 채팅방 Repository 단위테스트 리팩토링
